# openapi/basement-ui.yaml
openapi: 3.1.0
info:
  title: basement API
  version: 0.1.0
  description: 'Public OSS contract for basement''s API surface — the interface
      between

      the React frontend and Go backend. Frontend types are generated from this

      spec via openapi-typescript; backend handlers implement against it.


      **Versioning:** This is a 0.x spec (allowed to break). Contract freezes at

      version 1.0.0 when basement reaches feature-complete admin scope.

      '
  license:
      name: MIT
  contact:
      email: noreply@example.com
servers:
- url: http://localhost:8080/api/v1
  description: Local development server
...
paths:
...
  /auth/elevate/oidc/start:
      post:
        operationId: elevateOIDCStart
        summary: Begin OIDC step-up elevation (ADR-0003)
        description: |
          For OIDC-only users (no local password): mints a short-lived
          state token + returns the IdP's authorize URL with
          `prompt=login` (configurable via BASEMENT_OIDC_ELEVATION_PROMPT)
          and `max_age=0` so the IdP forces a FRESH re-authentication
          even if it has a cached session. The FE is expected to
          window.location.href to the returned redirect_url; the
          callback handler completes the elevation server-side.
        tags:
        - Auth
        x-basement-since: '1.2.0c'
        security:
        - cookieAuth: []
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required:
                - target_mode
                properties:
                  target_mode:
                    type: string
                    enum: [admin, elevated]
        responses:
          '200':
            description: Elevation flow started; FE must navigate to redirect_url.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    redirect_url:
                      type: string
                      format: uri
          '400':
            description: INVALID_TARGET_MODE / INVALID_JSON
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '401':
            description: UNAUTHORIZED
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '503':
            description: OIDC_NOT_CONFIGURED
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
      cookieAuth:
        type: apiKey
        in: cookie
        name: __Host-basement_session
        description: JWT session cookie issued after authentication. HttpOnly, Secure,
          SameSite=Strict.
  schemas:
      ErrorResponse:
        type: object
        required:
        - error
        properties:
          error:
            type: object
            required:
            - code
            - message
            properties:
              code:
                type: string
                example: DRIVER_UNSUPPORTED
              message:
                type: string
              details:
                type: object
                additionalProperties: true
...
security:
- cookieAuth: []